Improving test coverage and solving some issues

+++ src/main/java/japstudy/CompareAnswers.java
@@ -15,15 +15,15 @@ public final class CompareAnswers {
 
         List<String> set = bigrams(s);
         List<String> set2 = bigrams(s2);
-        double nx = set.size();
-        double ny = set2.size();
+        int nx = set.size();
+        int ny = set2.size();
         set2.removeAll(set);
 
 		if (nx + ny == 0) {
 			return 1;
 		}
 
-        return 1 - 2 * set2.size() / (nx + ny);
+        return 1. - 2. * set2.size() / (nx + ny);
 	}
 
 	private static List<String> bigrams(String s) {
